Django/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.19
djangorestframework==3.15.2
pillow==10.4.0
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
django-cors-headers==4.3.1
